Version arm

+++ pyasm4.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python3
+#!/usr/bin/env python
 
 import ctypes
 import numpy as np
@@ -8,7 +8,7 @@ arr = np.arange(100000000, dtype = np.int32)
 print(arr)
 
 dll = ctypes.cdll.LoadLibrary("./add4.so")
-dll.add4.restype = ctypes.c_long
+dll.add4.restype = ctypes.c_longlong
 dll.add4.argtypes = (
     np.ctypeslib.ndpointer(dtype=np.int32, ndim=1, flags='C_CONTIGUOUS'),
     ctypes.c_int
@@ -17,6 +17,6 @@ start = time()
 r = dll.add4(arr, len(arr))
 print(time() - start)
 start = time()
-s = np.sum(arr)
+s = np.sum(arr, dtype = np.int64)
 print(time() - start)
 print(r, s)
